a little cleanup to reduce compiler errors 5
