+++ src/board_constants.c
@@ -1,6 +1,7 @@
 //
 // Created by erena on 29.05.2024.
 //
+#pragma once
 
 #include "board_constants.h"
